local util refactor pt 2
